market-data/requirements.txt
websockets>=12.0
redis[hiredis]>=5.0
orjson>=3.9
structlog>=24.0
python-dotenv>=1.0
pytest>=8.0
pytest-asyncio>=0.23
fakeredis[aioredis]>=2.20
pytest-cov>=5.0
